scripts/pyproject.toml
[tool.poetry]
name = "openreview-backend"
version = "0.1.0"
description = ""
authors = ["Bastian Rieck <bastian.rieck@bsse.ethz.ch>"]

[tool.poetry.dependencies]
python = ">=3.8"
pandas = "^1.2.2"
openreview-py = "^1.0.20"
requests = "^2.25.1"
python-slugify = "^4.0.1"
filetype = "^1.0.7"
matplotlib = "^3.4.1"
wordcloud = "^1.8.1"
"pdfminer.six" = "^20201018"
python-dotenv = "^0.17.1"
spacy = "^3.4.1"
seaborn = "^0.12.0"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
